Take five exports with no consumer off the format's surface
architecture-first-principles C2(c): every new exported symbol read from the
consumer's chair, and five had no consumer outside the package.

`implicatedIn` and `implicating` were orphaned by the two commits before this
one — the first was a second ordering of the axis `blamed` already files in
path order, the second was `admits`' own scan. Both are deleted; the one test
that named `implicatedIn` asks the partition instead. `ADMITTED` is the shared
value `admits` answers with and is nobody's to name, so it stops being an
export. `withheld` is the validator's own step and has one in-package caller —
a public door for it would let anyone build a set whose `broken` names a path
the documents do not hold. And `findingsIn` — the directory's whole report as
one flat list — has no surface that wants it (a reader marks one file and opens
one page); it is what an ASSERTION wants, so it moves to `@olai/format/testlib`
beside `recordsOf`, which is where this package already keeps such a thing.

index.ts's own header says everything else in `src/` is internal; this makes
that true again. Behaviour unchanged — nothing outside the package could call
any of them.
