+++ packages/format/README.md
@@ -276,6 +276,6 @@ It also gives "absent" ONE spelling. An optional field holding nothing — `unde
 
 `src/failure.ts` is the other half of the error story: the four KINDS a write refuses with (`usage`, `not-found`, `validation`, `busy`), as schemas, so a refusal travels the wire and an MCP tool result as itself. Four classes rather than one with a `kind` string because they carry different things — only `validation` has a verdict to show, only `not-found` names what was missed — and a single struct would make every field optional and push "which fields are meaningful" back into prose. There were five: `derived` refused a write that would have stored a computed status, and went with derivation itself.
 
-`src/verdict.ts` is what a REFUSED WRITE answers with, and it stopped being a flat `ReadonlyArray<OutlineError>` on 2026-08-25 (the lowy-electricity debate's finding 2). The rows are still the product and still travel whole; what changed is that the questions consumers used to re-derive from them — and got wrong three different ways — are answered here, once. `implicating` / `implicatedBy` are which files a finding is about, which is `errors.ts`'s `Reach` axis read per finding rather than per code. `summaryOf(broken, n)` is a BOUNDED per-file face — state and counts, no rows — for a surface drawn over pages that are still live; the enumeration stays where a reader asked for it. It takes the entries `blamed` files and there is no second constructor over a verdict: the one caller that starts from one composes the two, so the step every reader makes — a verdict becomes per-file entries — stays visible at the call site. `admits(broken, files)` reads what is broken per FILE: `admitted` or `implicated(file, rows)`, with no whole-set member, so a refusal names its blocker rather than indicting a write that was innocent. It takes the entries `blamed` files, which is what makes it ONE question — a set carries them, and a caller holding a verdict takes the same step every other reader of one takes.
+`src/verdict.ts` is what a REFUSED WRITE answers with, and it stopped being a flat `ReadonlyArray<OutlineError>` on 2026-08-25 (the lowy-electricity debate's finding 2). The rows are still the product and still travel whole; what changed is that the questions consumers used to re-derive from them — and got wrong three different ways — are answered here, once. `implicatedBy` is which files ONE finding is about, which is `errors.ts`'s `Reach` axis read per finding rather than per code. `summaryOf(broken, n)` is a BOUNDED per-file face — state and counts, no rows — for a surface drawn over pages that are still live; the enumeration stays where a reader asked for it. It takes the entries `blamed` files and there is no second constructor over a verdict: the one caller that starts from one composes the two, so the step every reader makes — a verdict becomes per-file entries — stays visible at the call site. `admits(broken, files)` reads what is broken per FILE: `admitted` or `implicated(file, rows)`, with no whole-set member, so a refusal names its blocker rather than indicting a write that was innocent. It takes the entries `blamed` files, which is what makes it ONE question — a set carries them, and a caller holding a verdict takes the same step every other reader of one takes.
 
 And `blamed(report)` is what a LOAD does with a finding: the report filed under the files it BREAKS, as the per-file entries `set.ts` carries. It took the place of `tierOf`, a per-class table of what each finding cost a load, whose own prose said the missing half was the carry — a class moved off the whole-set blanket had to be CARRIED somewhere a reader could see it, or the finding would be true, consultable and invisible. The human ruled on 2026-08-29 that every finding is per file, so the shelf became a table with one row and this is the carry that replaced it (`verdict.test.ts` pins the ruling over generated broken sets: one file dark, its neighbours live and writable, both ends of a cross-file finding dark together).
