+++ packages/format/src/fixtures.testlib.ts
@@ -789,3 +789,38 @@ export const skippedAgenda = (derived: Derived, today: string): Agenda => {
     upcoming,
   }
 }
+
+/**
+ * EVERY ROW THE DIRECTORY HOLDS, once each, in presentation order — the report
+ * a validation reached, read back off the set that carries it.
+ *
+ * A set files the report under the files it breaks (`./verdict.ts`'s `blamed`),
+ * which is the shape every SURFACE wants: the sidebar marks a file, a page
+ * draws that file's rows, a banner counts them per file. This is the inverse,
+ * and it is exactly the report — `blamed` groups in path order and keeps each
+ * file's rows in the order the report held them, so flattening in that order
+ * and dropping the second sighting of a cross-file finding puts the list back
+ * the way `reportOf` sorted it.
+ *
+ * IT LIVES HERE AND NOT ON THE SET, because no surface asks it: a reader marks
+ * one file and opens one page, and the flat list is what an ASSERTION wants
+ * ("this corpus produces exactly these rows, in this order"). An export with no
+ * production caller is a promise the product does not make, and this package's
+ * own index says everything else in `src/` is internal.
+ *
+ * BY IDENTITY, which is what makes "the second sighting" a fact rather than a
+ * comparison: the two entries a cross-file finding lands in carry the same row
+ * object, because there is one finding.
+ */
+export const findingsIn = (set: OutlineSet): ReadonlyArray<OutlineError> => {
+  const seen = new Set<OutlineError>()
+  const rows: Array<OutlineError> = []
+  for (const entry of set.broken) {
+    for (const row of entry.errors) {
+      if (seen.has(row)) continue
+      seen.add(row)
+      rows.push(row)
+    }
+  }
+  return rows
+}
